html and css touchups
touched up the html and css. fixed sizing errors in the headings. created more space between headers and paragraphs. fixed the genre section background color issue so it is consistant with the rest of the background. Put a wrapper around contact and about since there's not much content -- I think it looks better. Hopefully when getting the api integrated the site looks more full. Still need to figure out what to do about a logo. Start working on JS and Jquery soon.