index, css, script updates
adding links with defer to header of html.

CSS - added styles for the reading list container of saved books and the display of the carousel. also styles for the save button. might need to look back and rearrange padding.

Script: wrote js and jquery for the carousel and the save to reading list functions. also save book function. need to test and debug more I think it's just about there. *initializing calls are commented out please uncomment these on next push*